Add fixture to return aioboto3 client
The asyncio client for boto3, aioboto3, takes the same arguments
as boto3 itself, but as a user of pytest-dynamodb I couldn't find
a simple and clean way to work out those arguments in my own code.
Adding an extra fixture upstream to return an aioboto3 client is
simpler and more reliable.

I briefly considered a different design where the boto3 session
class was parametrizable, but that would have been more complicated
because the setup and teardown code is slightly different between
boto3 and aioboto3.